pyproject.toml
[project]
name = "gym-assistant"
version = "0.1.0"
description = "Gym workout tracking assistant"
readme = "README.md"
requires-python = ">=3.12,<3.13"

dependencies = [
    "agate==1.9.1",
    "annotated-types==0.7.0",
    "asgiref==3.10.0",
    "attrs==25.4.0",
    "babel==2.17.0",
    "cachetools==6.2.2",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "colorama==0.4.6",
    "cryptography==46.0.3",
    "daff==1.4.2",
    "dbt==1.0.0.40.11",
    "dbt-adapters==1.20.1",
    "dbt-common==1.36.0",
    "dbt-core==1.10.15",
    "dbt-extractor==0.6.0",
    "dbt-postgres==1.9.1",
    "dbt-protos==1.0.397",
    "dbt-semantic-interfaces==0.9.0",
    "deepdiff==8.6.1",
    "dj-rest-auth==7.0.1",
    "Django==5.2.8",
    "django-allauth==65.13.0",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "dotenv==0.9.9",
    "drf-yasg==1.21.11",
    "google-api-core==2.28.1",
    "google-api-python-client==2.187.0",
    "google-auth==2.41.1",
    "google-auth-httplib2==0.2.1",
    "google-auth-oauthlib==1.2.3",
    "googleapis-common-protos==1.72.0",
    "greenlet==3.2.4",
    "gunicorn==23.0.0",
    "httplib2==0.31.0",
    "idna==3.11",
    "importlib-metadata==8.7.0",
    "inflection==0.5.1",
    "isodate==0.6.1",
    "Jinja2==3.1.6",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "leather==0.4.0",
    "MarkupSafe==3.0.3",
    "mashumaro==3.14",
    "more-itertools==10.8.0",
    "msgpack==1.1.2",
    "networkx==3.4.2",
    "oauthlib==3.3.1",
    "orderly-set==5.5.0",
    "packaging==25.0",
    "pandas==2.3.3",
    "parsedatetime==2.6",
    "pathspec==0.12.1",
    "proto-plus==1.26.1",
    "protobuf==6.33.1",
    "psycopg2-binary==2.9.11",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.23",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "PyJWT==2.10.1",
    "pyparsing==3.2.5",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "python-slugify==8.0.4",
    "pytimeparse==1.1.8",
    "pytz==2025.2",
    "PyYAML==6.0.3",
    "referencing==0.37.0",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "rpds-py==0.30.0",
    "rsa==4.9.1",
    "six==1.17.0",
    "snowplow-tracker==1.1.0",
    "SQLAlchemy==2.0.44",
    "sqlparse==0.5.3",
    "text-unidecode==1.3",
    "typing-inspection==0.4.2",
    "typing-extensions==4.15.0",
    "tzdata==2025.2",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "whitenoise==6.11.0",
    "zipp==3.23.0",
    # Linting & type checking
    "mypy>=1.14, <1.19",
    "ruff>=0.9",
    "django-stubs>=5.1",
    "pre-commit>=4.0",
    "pytest-playwright>=0.7.2",
    "playwright>=1.57.0",
    "alembic>=1.14",
    "pytest-django>=4.11.0",
    "pytest>=9.0.0",
    "sqlmodel>=0.0.31",
    "mailersend>=2.0.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "UP",  # pyupgrade (modern Python syntax)
]
ignore = [
    "E501",  # line too long (let formatter handle it)
]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true
warn_unused_ignores = true
strict = false  # not too annoying
exclude = [
    "infra/",
    "manage.py",
    "db/alembic/versions/",
]

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "backend.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "backend.settings"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--reuse-db"
